requirements.txt
altgraph==0.17.4
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
cryptography==42.0.5
defusedxml==0.7.1
Django==4.2.6
django-allauth==0.61.1
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
fastapi==0.108.0
Flask==3.0.2
h11==0.14.0
httptools==0.6.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.0
kombu==5.3.4
MarkupSafe==2.1.5
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pefile==2023.2.7
prompt-toolkit==3.0.43
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==2.0.0
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.32.0.post1
threadpoolctl==3.4.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.2.1
uvicorn==0.25.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
Werkzeug==3.0.1
